requirements.txt
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
coloredlogs==10.0
Deprecated==1.2.5
GitPython==2.1.11
humanfriendly==4.18
PyGithub==1.43.7
requests==2.22.0
tqdm==4.32.1
urllib3==1.25.3
